requirements.txt
paho-mqtt
pandas
pika